v1.1.2 merge

+++ CrossApp/ccTypeInfo.h
@@ -19,7 +19,7 @@ public:
 
 static inline unsigned int getHashCodeByString(const char *key)
 {
-	unsigned int len = strlen(key);
+	unsigned int len = (unsigned int)strlen(key);
 	const char *end=key+len;
 	unsigned int hash;
